@@ -11,17 +11,17 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:quick_blue_example/main.dart';
 
 void main() {
-  testWidgets('Verify Platform version', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+  // testWidgets('Verify Platform version', (WidgetTester tester) async {
+  //   // Build our app and trigger a frame.
+  //   await tester.pumpWidget(MyApp());
 
-    // Verify that platform version is retrieved.
-    expect(
-      find.byWidgetPredicate(
-        (Widget widget) => widget is Text &&
-                           widget.data.startsWith('Running on:'),
-      ),
-      findsOneWidget,
-    );
-  });
+  //   // Verify that platform version is retrieved.
+  //   expect(
+  //     find.byWidgetPredicate(
+  //       (Widget widget) => widget is Text &&
+  //                          widget.data.startsWith('Running on:'),
+  //     ),
+  //     findsOneWidget,
+  //   );
+  // });
 }
